src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile src/requirements.in
#
-e file:src/vehicle_model
    # via -r src/requirements.in
aiohttp==3.8.1
    # via dapr
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
charset-normalizer==2.0.12
    # via aiohttp
cloudevents==1.2.0
    # via -r src/requirements.in
dapr==1.6.0
    # via -r src/requirements.in
deprecation==2.1.0
    # via cloudevents
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
grpcio==1.44.0
    # via
    #   -r src/requirements.in
    #   dapr
idna==3.3
    # via yarl
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
packaging==21.3
    # via deprecation
protobuf==3.19.4
    # via
    #   -r src/requirements.in
    #   dapr
pyparsing==3.0.8
    # via packaging
python-dateutil==2.8.2
    # via dapr
six==1.16.0
    # via
    #   grpcio
    #   python-dateutil
yarl==1.7.2
    # via aiohttp
